requirements.txt
requests  
matplotlib  
python-dotenv  
textblob  
APScheduler